+++ src/base/b-list/b-list.ts
@@ -687,6 +687,15 @@ export default class bList extends iData implements iVisible, iWidth, iItems {
 		return iItems.getItemKey(this, item, i);
 	}
 
+	/**
+	 * Returns href for item. Is used at the rendering stage
+	 *
+	 * @param item
+	 */
+	protected getHref(item: Item): CanUndef<string> {
+		return item.href;
+	}
+
 	protected override initModEvents(): void {
 		super.initModEvents();
 		iVisible.initModEvents(this);